kolesa/requirements.txt
playwright
pandas
apache-airflow
tqdm
sqlalchemy